+++ src/components/GroupResult.jsx
